Test Bench: Player Movement & key events
