+++ mlx_test_bench/main.c
@@ -1,159 +1,63 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-#include "../inc/macos_keycodes.h"
+#include "bench.h"
 #include "chong_sketch.h"
 
-#include <time.h>
-#include <unistd.h>
+t_bench	bench;
 
-# define WIDTH 1280
-# define HEIGHT 720
-
-typedef	struct s_bench
+void	exit_test_bench(void *arg)
 {
-	void	*mlx;
-	void	*window;
-}	t_bench;
+	(void)arg;
 
-int	key_press(int keycode, t_bench *bench)
-{
-	if (keycode == KEY_ESC)
-	{
-		csketch_gc();
-		mlx_destroy_window(bench->mlx, bench->window);
-		exit(0);
-	}
-	return (keycode);
+	mlx_destroy_window(bench.mlx.mlx_ptr, bench.mlx.win_ptr);
+	printf("----------------------------------------\n");
+	printf("Exiting Test Bench...\n");
+	exit(0);
 }
 
-# define GRID_W 10
-# define GRID_H 7
-
-typedef struct s_vec_2_int
+// Could be a much more efficient way to handle key press.
+// Iterative approach will be fine as there is only 6 possible keystrokes that
+// are valid in mandatory cub3d.
+static int	handle_keypress(int keycode)
 {
-	int	x;
-	int	y;
-}	t_v2i;
+	t_key_event	*events;
 
-typedef struct s_vertex
-{
-	t_v2i		position;
-	uint32_t	colour;
-}	t_vertex;
-
-size_t	ft_strlen(const char *s)
-{
-	size_t	i;
-
-	if (!s)
-		return (0);
-	i = 0;
-	while (s[i])
-		i++;
-	return (i);
+	events = bench.events.key_press;
+	for (size_t i = 0; i < SUPPORTED_KEYCODE_COUNT; i++)
+	{
+		if (keycode == events[i].code)
+		{
+			events[i].handler();
+			return (0);
+		}
+	}
+	printf("Key press %i is not supported.\n", keycode);
+	return (keycode);
 }
 
 int	main(void)
 {
-	t_bench	bench;
+	t_mlx	*mlx;
 
-	bench.mlx = mlx_init();
-	if (!bench.mlx)
+	mlx = &bench.mlx;
+	mlx->mlx_ptr = mlx_init();
+	if (!mlx->mlx_ptr)
 	{
-		fprintf(stderr, "Error: Could not initialise mlx pointer.\n");
+		fprintf(stderr, "Error\nCould not initialise mlx pointer.\n");
 		return (1);
 	}
-	bench.window = mlx_new_window(bench.mlx, WIDTH, HEIGHT, "Chong Sketch: Test Bench");
-	if (!bench.window)
+	mlx->win_ptr = mlx_new_window(mlx->mlx_ptr, WIN_WIDTH, WIN_HEIGHT, WIN_TITLE);
+	if (!mlx->win_ptr)
 	{
-		fprintf(stderr, "Error: Could not initialise mlx pointer.\n");
+		fprintf(stderr, "Error\nCould not create a new mlx window instance.\n");
 		return (1);
 	}
-	// mlx_key_hook(bench.window, key_press, &bench);
 
-	t_csketch_errno err = csketch_init(bench.mlx, bench.window, WIDTH, HEIGHT);
-	if (err)
-	{
-		fprintf(stderr, "Error: Could not initialise chong sketch canvas: %s\n", csketch_errno(err));
-		return (err);
-	}
+	t_key_event	*keypress;
+	keypress = bench.events.key_press;
+	keypress[I_ESC].code = KEY_ESC;
+	keypress[I_ESC].handler = exit_test_bench;
 
-	// csketch_fill(WIDTH, HEIGHT, 0x2B2B2B);
-	// csketch_rectangle(600, 400, WIDTH/2 - 600/2, HEIGHT/2 - 400/2, 0xFFFFFF);
-	// csketch_circle(WIDTH/2, HEIGHT/2, 120, 0xFF0000);
-	
-	// csketch_line(100, 100, 700, 500, 0x20FF20);	// Works...
-	// csketch_line(700-1, 100, 100-1, 500, 0x20FF20);	// Works...
-	// csketch_line(WIDTH/2, 100, WIDTH/2, 500, 0x20FF20);	// Works...
-	// csketch_line(100, HEIGHT/2, 700, HEIGHT/2, 0x20FF20);	// Works...
+	mlx_hook(mlx->win_ptr, 2, 0, handle_keypress, NULL);
 
-	char	grid[GRID_H][GRID_W] = {
-		{'1','1','1','1','1','1','1','1','1','1'},
-		{'1','0','0','0','0','1','0','0','0','1'},
-		{'1','0','0','1','0','0','0','0','0','1'},
-		{'1','0','0','1','0','0','0','0','0','1'},
-		{'1','0','0','0','0','1','0','0','0','1'},
-		{'1','0','0','0','0','1','0','0','0','1'},
-		{'1','1','1','1','1','1','1','1','1','1'},
-	};
-
-	size_t tile_size = 64;
-	t_vertex vertices[GRID_W * GRID_H];
-
-	for (size_t	y = 0; y < GRID_H; y++)
-	{
-		for (size_t x = 0; x < GRID_W; x++)
-		{
-			t_vertex *v = &vertices[y * GRID_W + x];
-			v->position.x = x * tile_size;
-			v->position.y = y * tile_size;
-			if (grid[y][x] == '1')
-				v->colour = 0x4848A8;
-			else
-				v->colour = 0xBFBFBF;
-		}
-	}
-
-	// Render Loop.
-	const char *htag = "Rendering...\t";
-	write(STDOUT_FILENO, htag, ft_strlen(htag));
-	clock_t time = clock();
-
-	csketch_clear();
-	csketch_fill(WIDTH, HEIGHT, 0x2B2B2B);
-	for (size_t y = 0; y < GRID_H; y++)
-	{
-		for (size_t x = 0; x < GRID_W; x++)
-		{
-			t_vertex v = vertices[y * GRID_W + x];
-			t_v2i p = v.position;
-			p.x += WIDTH/2 - (GRID_W*tile_size)/2;
-			p.y += HEIGHT/2 - (GRID_H*tile_size)/2;
-			csketch_rectangle(tile_size, tile_size, p.x, p.y, v.colour);
-			csketch_line(p.x, p.y, p.x + tile_size, p.y, 0);
-			csketch_line(p.x, p.y, p.x, p.y + tile_size, 0);
-			if (y == GRID_H - 1)
-				csketch_line(p.x, p.y + tile_size, p.x + tile_size, p.y + tile_size, 0);
-			if (x == GRID_W - 1)
-				csketch_line(p.x + tile_size, p.y, p.x + tile_size, p.y + tile_size, 0);
-		}
-	}
-
-	time = clock() - time;
-	double elapsed = (((double)time)/CLOCKS_PER_SEC)*1000;
-	fprintf(stdout, "Elapsed Time: %.3fms\n", elapsed);
-	fprintf(stdout, "FPS: %d\n", (int)(1000/elapsed));
-
-	err = csketch_outppm("grid.ppm", WIDTH, HEIGHT);
-	if (err)
-	{
-		fprintf(stderr, "Error: Could not write to ppm file \"grid.ppm\": %s\n", csketch_errno(err));
-		return (err);
-	}
-	mlx_destroy_image(bench.mlx, final_csketch());
-	csketch_gc();
-	// mlx_put_image_to_window(bench.mlx, bench.window, final_csketch(), 0, 0);
-	// mlx_loop(bench.mlx);
+	mlx_loop(mlx->mlx_ptr);
 	return (0);
 }
